requirements.txt
streamlit
pandas
plotly
sqlalchemy
langchain
langchain-openai
python-dotenv
scikit-learn
pydantic
